feat: added pause/resume to main menu
This seems to have broken a bunch of my tests. I spent some time trying to get them to work but had to give up due to time constraints. The next step with the tests is probably to revert to the last commit where they seemed to be working well and see exactly which change causes them to break.